Revert "move develop class out of setup.py"
This reverts commit 1b402daf1b81cb895c45b543d8314fa585fe3b7f. It creates
issues when installing using e.g. poetry, and I think I prefer this
small class definition above a sys path hack.
