+++ setup.py
@@ -2,13 +2,12 @@ import os
 import shlex
 from glob import glob
 
-import numpy
-from Cython.Distutils import build_ext
 from setuptools import setup
 from setuptools.command.build_clib import build_clib
+from setuptools.command.develop import develop as st_develop
 from setuptools.extension import Extension
-
-from ms2pip.distutils.develop import develop
+from Cython.Distutils import build_ext
+import numpy
 
 VERSION = "3.6.2"
 
@@ -79,6 +78,14 @@ extensions = [
 ]
 
 
+class develop(st_develop):
+    __doc__ = st_develop.__doc__
+
+    def install_for_development(self):
+        self.run_command('build_clib')
+        st_develop.install_for_development(self)
+
+
 setup(
     name=NAME,
     version=VERSION,
@@ -105,5 +112,6 @@ setup(
     libraries=[libms2pip],
     ext_modules=extensions,
     include_dirs=[numpy.get_include()],
+    # library_dirs=[os.path.join(os.path.dirname(__file__), 'ms2pip', 'cython_modules')],
     cmdclass={"build_clib": build_clib, "build_ext": build_ext, 'develop': develop},
 )
